+++ src/spellbot/errors.py
@@ -14,6 +14,11 @@ class AdminOnlyError(SpellBotError):
         super().__init__(message or "This command is only available to SpellBot admins.")
 
 
+class GuildOnlyError(SpellBotError):
+    def __init__(self, message: Optional[str] = None):
+        super().__init__(message or "This command is only available within a guild.")
+
+
 class UserBannedError(SpellBotError):
     def __init__(self, message: Optional[str] = None):
         super().__init__(message or "This user has been banned from using SpellBot.")